+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Models: throwing `Add()` method to builders in addition to non-throwing `TryAdd()`.
 
+### Changed
+- Traversal: The order of parameters in `Enumerate*()` signatures for `Bfs<>` and `Dfs<>`.
+
 ### Removed
 - Models: `GraphBuilderExtensions` class with `TryAdd<>()` extension method.